// shared/types/server/kpCreateDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Omnition Api Documentation
 * Документация Api для Omnition
 * OpenAPI spec version: 1.0
 */
import type { FileJson } from "./fileJson";

export interface KpCreateDto {
    /** Название КП */
    name: string;
    /** Номер КП */
    numberKp: string;
    /** Дата КП */
    date: string;
    /** Имя клиента */
    clientName: string;
    /** Номер заказа */
    numberOrder: string;
    /** Адрес заказа */
    orderAddress: string;
    /** Контактное лицо */
    contactPerson: string;
    /** Детали КП */
    details: string;
    /** Смета на работы */
    estimate: FileJson;
    /** Документы КП */
    documents: FileJson[];
    /** Оплата и сроки работ */
    payAndDeadlines: string[];
    /** Условия для производства работ */
    productionConditions: string[];
    /** Продукты для любых поверхностей */
    anySurface: boolean;
    /** Преимущества архитектурного бетона */
    advantagesOfConcrete: boolean;
    /** Варианты механизации */
    variationsOfTheMechanism: boolean;
    /** Шлифовка */
    grinding: boolean;
    /** Собственный передвижной завод */
    ownMobileFactory: boolean;
    /** Варианты интерьера */
    interiorOptions: boolean;
    /** Лестницы */
    stairs: boolean;
    /** Благоустройство */
    landscaping: boolean;
    /** Ремонт и восстановление */
    repairAndRestoration: boolean;
    /** ID менеджера */
    managerId: number;
    /** ID дизайна узла */
    nodeDesignId: number;
    /** ID клиента */
    clientId: number;
}